+++ components/core/src/clp/ffi/ir_stream/search/QueryHandlerReq.hpp
@@ -1,0 +1,43 @@
+#ifndef CLP_FFI_IR_STREAM_SEARCH_QUERYHANDLERREQ_HPP
+#define CLP_FFI_IR_STREAM_SEARCH_QUERYHANDLERREQ_HPP
+
+#include <type_traits>
+
+#include "NewProjectedSchemaTreeNodeCallbackReq.hpp"
+#include "QueryHandler.hpp"
+
+namespace clp::ffi::ir_stream::search {
+/**
+ * Defines an empty query handler that can be used with `clp::ffi::ir_stream::Deserializer` to
+ * deserialize an IR stream without performing any query evaluation.
+ */
+struct EmptyQueryHandler {};
+
+/**
+ * A type trait to determine if a given type is a valid `clp::ffi::ir_stream::search::QueryHandler`.
+ * @tparam T The type to check.
+ */
+template <typename T>
+struct IsQueryHandler : std::false_type {};
+
+/**
+ * Specialization of `IsQueryHandler` for `clp::ffi::ir_stream::search::QueryHandler`.
+ * This specialization evaluates to `true` for valid `QueryHandler` types.
+ * @tparam NewProjectedSchemaTreeNodeCallbackType
+ */
+template <NewProjectedSchemaTreeNodeCallbackReq NewProjectedSchemaTreeNodeCallbackType>
+struct IsQueryHandler<QueryHandler<NewProjectedSchemaTreeNodeCallbackType>> : std::true_type {};
+
+/**
+ * Requirements for a query handler that can be used with`clp::ffi::ir_stream::Deserializer`.
+ * A valid query handler must either be `EmptyQueryHandler` or satisfy the `IsQueryHandler` trait,
+ * and must also be move-constructible.
+ * @tparam QueryHandlerType The type to validate as a query handler.
+ */
+template <typename QueryHandlerType>
+concept QueryHandlerReq = (std::is_same_v<QueryHandlerType, EmptyQueryHandler>
+                           || IsQueryHandler<QueryHandlerType>::value)
+                          && std::is_move_constructible_v<QueryHandlerType>;
+}  // namespace clp::ffi::ir_stream::search
+
+#endif  // CLP_FFI_IR_STREAM_SEARCH_QUERYHANDLERREQ_HPP
